LayoutTests/accessibility/mac/static-text-with-aria-label-expected.txt

"
This tests that aria-label on a role='text' does not duplicate the AXValue in the AXDescription

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS accessibilityController.focusedElement.description is 'AXDescription: '
PASS accessibilityController.focusedElement.stringValue is 'AXValue: inch'
PASS successfullyParsed is true

TEST COMPLETE
